swap HydroForces/HydroSystem naming for semantic clarity
Rename user-facing façade from HydroForces to HydroSystem to align with
Chrono conventions (ChSystem = main entry point). Rename internal
Chrono-free force engine from HydroSystem to HydroForces (in
hydrochrono::hydro namespace) since it computes forces.

Changes:
- HydroForces → HydroSystem (user-facing façade, include/hydroc/hydro_system.h)
- HydroSystem → hydrochrono::hydro::HydroForces (internal engine, include/hydroc/core/hydro_forces.h)
- Add backward-compat aliases: TestHydro, HydroForces → HydroSystem
- Add compatibility header (hydro_forces.h → hydro_system.h) for existing includes
- Update ChronoHydroCoupler, setup_from_yaml, CMakeLists.txt

No behavioral changes; all existing tests pass.
