+++ CMakeLists.txt
@@ -345,7 +345,7 @@ set(HYDROCHRONO_SOURCES
     src/hydro/config/setup_from_yaml.cpp
     src/hydro/config/yaml_discovery.cpp
     # Core facade and utilities
-    src/hydro/hydro_forces.cpp
+    src/hydro/hydro_system.cpp
     src/hydro/utils/helper.cpp
     # Hydro modules
     src/hydro/waves/wave_base.cpp
@@ -354,7 +354,7 @@ set(HYDROCHRONO_SOURCES
     src/hydro/waves/irregular_wave.cpp
     src/hydro/radiation/radiation_rirf_processing.cpp
     src/hydro/radiation/radiation_rirf_convolution.cpp
-    src/hydro/core/hydro_system.cpp
+    src/hydro/core/hydro_forces.cpp
     # Chrono coupling layer (bridges Chrono physics with Chrono-free core)
     src/hydro/chrono/chrono_state_utils.cpp
     src/hydro/chrono/chrono_hydro_coupler.cpp
