+++ include/hydroc/hydro_forces.h
@@ -1,524 +1,19 @@
 /*********************************************************************
  * @file  hydro_forces.h
- * @brief HydroForces façade: attaches hydrodynamic forces to Chrono bodies.
+ * @brief Backward-compatibility header - includes hydro_system.h
  *
- * MAIN TYPES:
- *   - HydroForces: Primary hydrodynamics façade (adapter over HydroSystem)
- *   - ForceFunc6d: Wraps 6-DOF force/torque callbacks per body
- *   - ComponentFunc: Per-DOF force function for Chrono's ChForce system
+ * This header is retained for backward compatibility. It simply includes
+ * the new hydro_system.h header which defines HydroSystem (the user-facing
+ * hydrodynamics façade) and the HydroForces alias.
  *
- * ROLE: This is the main entry point for C++ usage. Internally delegates
- * to HydroSystem (Chrono-free core) and ChronoHydroCoupler. Used by both
- * C++ tests/demos and the YAML runner (hydrochrono.exe).
- *
- * KEY ASSUMPTIONS: 6 DOF per body, bodies named "body1", "body2", etc.
+ * @deprecated Include <hydroc/hydro_system.h> directly in new code.
  *********************************************************************/
 
 #ifndef HYDRO_FORCES_H
 #define HYDRO_FORCES_H
 
-// Include hydro_types.h FIRST to ensure BodyForces and GeneralizedForce are available
-// before any other includes that might conflict (e.g., config/hydro_config.h)
-#include <hydroc/core/hydro_types.h>
+// Forward to the new header
+#include <hydroc/hydro_system.h>
 
-// Standard includes
-#include <cstdio>
-#include <filesystem>
-#include <limits>
-#include <memory>
-#include <vector>
+#endif  // HYDRO_FORCES_H
 
-// Eigen includes (for GeneralizedForce and BodyForces)
-#include <Eigen/Dense>
-
-// Chrono library includes
-// Note: We include only the headers strictly necessary for the types used in this header.
-// Implementation-only Chrono headers are included in hydro_forces.cpp.
-// ChForce.h includes ChFunction internally, so we don't need a separate include.
-#include <chrono/physics/ChBody.h>
-#include <chrono/physics/ChForce.h>
-#include <chrono/physics/ChLoadContainer.h>
-
-// Hydroc library includes
-#include <hydroc/io/h5_reader.h>
-#include <hydroc/waves/wave_base.h>
-#include <hydroc/waves/regular_wave.h>
-#include <hydroc/waves/irregular_wave.h>
-
-// Public hydroc includes (for SystemState)
-#include <hydroc/core/system_state.h>
-
-// Radiation types (canonical definitions - single source of truth)
-#include <hydroc/radiation/radiation_types.h>
-
-namespace hydrochrono::hydro {
-// Forward declarations
-class HydrostaticsComponent;
-class RadiationComponent;
-class ExcitationComponent;
-class HydroSystem;
-class ChronoHydroCoupler;
-// Types GeneralizedForce and BodyForces are defined in hydro_types.h (included above)
-}
-
-class ForceFunc6d;
-class HydroForces;
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Degrees of Freedom Constant
-// ─────────────────────────────────────────────────────────────────────────────
-/// Number of degrees of freedom per rigid body: surge, sway, heave, roll, pitch, yaw.
-constexpr int kDofPerBody = 6;
-
-/**
- * @brief Per-DOF force function for Chrono's ChForce system.
- *
- * Inherits from chrono::ChFunction to provide time-dependent force values.
- * Each instance represents a single degree of freedom (surge, sway, heave, roll, pitch, yaw).
- */
-class ComponentFunc : public chrono::ChFunction {
-  public:
-    /**
-     * @brief Default constructor.
-     * @note Do not use this constructor.
-     */
-    ComponentFunc();
-
-    /**
-     * @brief Copy constructor.
-     */
-    ComponentFunc(const ComponentFunc& old);
-
-    /**
-     * @brief Construct a new ComponentFunc object.
-     *
-     * @param b Pointer to ForceFunc6d member object.
-     * @param i Index for which component this ComponentFunc object refers to. Corresponds
-     * to the force's degree of freedom, where:
-     * (0,1,2,3,4,5) -> (surge, sway, heave, roll, pitch, yaw).
-     */
-    ComponentFunc(ForceFunc6d* b, int i);
-
-    /**
-     * @brief Clone the object.
-     * @note This is a required override since ComponentFunc inherits from ChFunction.
-     * @return A cloned ComponentFunc object.
-     */
-    virtual ComponentFunc* Clone() const override;
-
-    /**
-     * @brief Calculate the force value for a specific time.
-     *
-     * @param x Time from simulation.
-     * @return Force on body in the specified degree of freedom at time x.
-     */
-    virtual double GetVal(double x) const override;
-
-  private:
-    ForceFunc6d* base_;  ///< Pointer to the full 6D force on the body.
-    int index_;          ///< Index representing force degree of freedom on the body.
-};
-
-/**
- * @brief Organizes the functional (time-dependent) forces in each degree of freedom (6 total) for a body.
- */
-class ForceFunc6d {
-  public:
-    /**
-     * @brief Initializes an array of ComponentFunc objects and pointers to each force/torque.
-     */
-    ForceFunc6d();
-
-    /**
-     * @brief Initializes hydro force info from H5FileInfo and the ChBody this force will be applied to.
-     *
-     * @param object The body in the system to which this 6-dimensional force is being applied.
-     * @param all_hydro_forces_user The HydroForces class where the total force on all bodies is calculated.
-     */
-    ForceFunc6d(std::shared_ptr<chrono::ChBody> object, HydroForces* all_hydro_forces_user);
-
-    /**
-     * @brief Copy constructor that ensures the force is only added to a body once.
-     *
-     * @note Avoid using the default copy constructor. ApplyForceAndTorqueToBody() should only ever be applied once.
-     *
-     * @param old ForceFunc6d object to copy from.
-     */
-    ForceFunc6d(const ForceFunc6d& old);
-
-    /**
-     * @brief Calculates the force on a given degree of freedom.
-     *
-     * @param i Index corresponding to the degree of freedom. Assumed to be 0-5 only.
-     * @return Value of force in the i-th degree of freedom for the associated body.
-     */
-    double CoordinateFunc(int i);
-
-  private:
-    /**
-     * @brief Initializes force components.
-     */
-    void SetForce();
-
-    /**
-     * @brief Initializes torque components.
-     */
-    void SetTorque();
-
-    /**
-     * @brief Adds this force to the body's list of applied forces.
-     *
-     * @warning Ensure this function is not called multiple times for the same force.
-     */
-    void ApplyForceAndTorqueToBody();
-
-    std::shared_ptr<chrono::ChBody> body_;          ///< Pointer to the body this force is applied to.
-    
-    /// Body index in the system (1-indexed, parsed from Chrono body name "bodyN").
-    /// This matches Chrono's body naming convention where bodies are named body1, body2, etc.
-    int b_num_;
-    
-    ComponentFunc forces_[kDofPerBody];                       ///< Forces for each degree of freedom.
-    std::shared_ptr<ComponentFunc> force_ptrs_[kDofPerBody];  ///< Pointers to the forces.
-    std::shared_ptr<chrono::ChForce> chrono_force_;  ///< Chrono force for the body.
-    std::shared_ptr<chrono::ChForce> chrono_torque_; ///< Chrono torque for the body.
-    HydroForces* all_hydro_forces_;                 ///< Pointer to HydroForces for calculations.
-};
-
-// Forward declaration of ChLoadAddedMass (defined in added_mass.h, included in .cpp)
-class ChLoadAddedMass;
-
-// Lightweight hydrodynamics profiling stats
-struct HydroProfileStats {
-    double hydrostatics_seconds = 0.0;
-    double radiation_seconds    = 0.0;
-    double waves_seconds        = 0.0;
-    int hydrostatics_calls      = 0;
-    int radiation_calls         = 0;
-    int waves_calls             = 0;
-};
-
-// ═══════════════════════════════════════════════════════════════════════════════
-//
-//  HydroForces — Primary Hydrodynamics Façade / Adapter
-//
-// ═══════════════════════════════════════════════════════════════════════════════
-/**
- * @brief Primary hydrodynamics façade for HydroChrono applications.
- *
- * HydroForces is the main user-facing object for attaching hydrodynamic forces to
- * Chrono bodies. This is the production hydrodynamics adapter used by all workflows.
- *
- * @note The historical name "TestHydro" is retained as an alias for backward
- *       compatibility. New code should use `HydroForces`.
- *
- * ## What It Does
- *
- * HydroForces bridges three layers:
- *   1. **Chrono bodies** — The physical bodies in the Chrono simulation.
- *   2. **HydroSystem (core)** — Chrono-free hydrodynamic force computation.
- *   3. **Chrono force callbacks** — ChForce functions that Chrono calls each timestep.
- *
- * ## Typical Usage
- *
- * @code{.cpp}
- *   // Create Chrono bodies
- *   auto body1 = chrono_types::make_shared<ChBody>();
- *   body1->SetName("body1");
- *   system.Add(body1);
- *
- *   // Attach hydrodynamic forces (with or without waves)
- *   std::vector<std::shared_ptr<ChBody>> bodies = {body1};
- *   HydroForces hydro(bodies, "path/to/hydro_data.h5");
- *   hydro.AddWaves(std::make_shared<RegularWave>(...));
- *
- *   // Run simulation — Chrono automatically queries HydroForces for forces
- * @endcode
- *
- * ## Used By
- *
- *   - **C++ tests and demos** — Directly instantiate HydroForces with bodies and H5 file.
- *   - **YAML runner (hydrochrono.exe)** — SetupHydroFromYAML creates HydroForces internally.
- *
- * ## Architecture Notes
- *
- *   - Internally delegates to HydroSystem + ChronoHydroCoupler for force computation.
- *   - Maintains velocity history for radiation damping convolution.
- *   - Caches forces per timestep to avoid redundant computation.
- *
- * ## Body Indexing Convention
- *
- *   Chrono bodies must be named "body1", "body2", etc. The numeric suffix is parsed
- *   to determine body order. Internally, body indices are **1-based** in the callback
- *   interface (CoordinateFuncForBody) to match this naming convention.
- *
- * @see TestHydro (backwards-compatibility alias)
- * @see HydroSystem (Chrono-free hydrodynamic core)
- * @see ChronoHydroCoupler (extracts state from Chrono bodies)
- */
-class HydroForces {
-  public:
-    HydroForces() = delete;
-
-    /**
-     * @brief Main constructor for initializing the HydroForces class.
-     *
-     * Sets up vector of bodies, h5 file info, and hydro inputs. If no waves are given,
-     * this constructor defaults to using NoWave.
-     *
-     * @param user_bodies List of pointers to bodies for the hydro forces.
-     * @param h5_file_name Name of the h5 file where hydro data is stored.
-     * @param waves WaveBase object. Defaults to NoWave if not provided.
-     */
-    HydroForces(std::vector<std::shared_ptr<chrono::ChBody>> user_bodies,
-                std::string h5_file_name,
-                std::shared_ptr<WaveBase> waves = std::make_shared<NoWave>());
-
-    // Destructor (defined in .cpp to allow unique_ptr to incomplete type)
-    ~HydroForces();
-
-    // Deleted copy constructor and assignment operator for safety.
-    HydroForces(const HydroForces& old) = delete;
-    HydroForces& operator=(const HydroForces& rhs) = delete;
-
-    /**
-     * @brief Adds waves class to force calculations depending on if regular or irregular waves.
-     *
-     * Also initializes h5 data for wave force class.
-     *
-     * @param waves the specific WaveBase class to add to the system
-     */
-    void AddWaves(std::shared_ptr<WaveBase> waves);
-
-    /**
-     * @brief Legacy API: Computes hydrostatic forces for a 6N dimensional system.
-     *
-     * NOTE: Retained for backward compatibility. The main force path now uses HydroSystem
-     * (via CoordinateFuncForBody). This method is not called during normal simulation.
-     *
-     * @return 6N dimensional force for 6 DOF and N bodies in system.
-     */
-    std::vector<double> ComputeForceHydrostatics();
-
-    /**
-     * @brief Legacy API: Computes radiation damping forces for a 6N dimensional system.
-     *
-     * NOTE: Retained for backward compatibility. The main force path now uses HydroSystem
-     * (via CoordinateFuncForBody). This method is not called during normal simulation.
-     *
-     * Returns positive damping magnitudes (legacy convention); the main path applies
-     * the correct sign internally via RadiationComponent.
-     *
-     * @return 6N dimensional force for 6 DOF and N bodies in system.
-     */
-    std::vector<double> ComputeForceRadiationDampingConv();
-
-    /**
-     * @brief Legacy API: Computes wave excitation forces for a 6N dimensional system.
-     *
-     * NOTE: Retained for backward compatibility. The main force path now uses HydroSystem
-     * (via CoordinateFuncForBody). This method is not called during normal simulation.
-     *
-     * @return 6N dimensional force for 6 DOF and N bodies in system.
-     */
-    Eigen::VectorXd ComputeForceWaves();
-    // Expose the wave object (read-only) so callers can query inputs if needed
-    std::shared_ptr<WaveBase> GetWave() const { return user_waves_; }
-
-    /**
-     * @brief Fetches the RIRF value from the h5 file based on the provided indices.
-     *
-     * @param row Index representing the body number and DOF index [0,...,5,...6N-1] for rows of RIRF.
-     * @param col Column index in RIRF matrix [0,...,5,...6N-1].
-     * @param st Index representing the timestep in RIRF, usually in the range [0,...1000].
-     *
-     * @return The value from the h5 file RIRF matrix.
-     */
-    double GetRIRFval(int row, int col, int st);
-
-    // ─────────────────────────────────────────────────────────────────────────
-    // Radiation Convolution Configuration
-    // ─────────────────────────────────────────────────────────────────────────
-    // These types are defined in the radiation module (single source of truth):
-    //   - hydrochrono::hydro::RadiationConvolutionMode (enum)
-    //   - hydrochrono::hydro::TaperedDirectOptions (struct)
-    // See: src/hydro/force_components/radiation_component.h
-    //      src/hydro/radiation/radiation_rirf_processing.h
-
-    /**
-     * @brief Set the radiation convolution mode. Default is Baseline.
-     * @param mode RadiationConvolutionMode::Baseline or ::TaperedDirect
-     */
-    void SetRadiationConvolutionMode(hydrochrono::hydro::RadiationConvolutionMode mode);
-
-    /**
-     * @brief Set options for TaperedDirect preprocessing.
-     * @param opts TaperedDirectOptions struct with smoothing, tapering, and export settings
-     */
-    void SetTaperedDirectOptions(const hydrochrono::hydro::TaperedDirectOptions& opts);
-
-    /**
-     * @brief Set the directory where diagnostics (e.g., CSVs) should be written.
-     */
-    void SetDiagnosticsOutputDirectory(const std::string& dir) {
-        diagnostics_output_dir_ = dir;
-        InvalidateRadiationComponent();  // Invalidate component to recreate with new settings
-    }
-
-    /**
-     * @brief Calculates or retrieves the total force on a specific body in a particular degree of freedom.
-     *
-     * This is the main entry point called by Chrono's ChForce callbacks during simulation.
-     * Forces are computed once per timestep and cached; subsequent calls within the same
-     * timestep return the cached value.
-     *
-     * @note **1-Based Body Indexing**: The body index `b` is 1-based (body1 → b=1, body2 → b=2, etc.)
-     *       to match Chrono's body naming convention. This is intentional and matches how
-     *       ForceFunc6d parses body names like "body1", "body2", etc.
-     *
-     * @param b Body index (1-based: valid range is [1, num_bodies]).
-     * @param i Degree of Freedom (DOF) index (0-based: 0=surge, 1=sway, 2=heave, 3=roll, 4=pitch, 5=yaw).
-     *
-     * @return Force component for body 'b' in DOF 'i' (Newtons for forces, Newton-meters for torques).
-     *
-     * @throws std::out_of_range if b or i are out of valid range.
-     */
-    double CoordinateFuncForBody(int b, int i);
-
-    // Hydrodynamics profiling accessors
-    HydroProfileStats GetProfileStats() const { return profile_stats_; }
-
-    /**
-     * @brief Enable or disable profiling in HydroSystem.
-     * 
-     * When enabled, HydroSystem measures timing for each force component.
-     * Disabled by default to avoid overhead in normal runs.
-     * 
-     * @param enabled True to enable profiling, false to disable
-     */
-    void SetProfilingEnabled(bool enabled);
-
-    // Compare mode: legacy debugging feature, retained for API compatibility.
-    // No longer affects behavior since the main path now uses HydroSystem exclusively.
-    void SetCompareMode(bool enable) { compare_mode_ = enable; }
-    bool GetCompareMode() const { return compare_mode_; }
-
-  private:
-    // Class properties related to the body and hydrodynamics
-    std::vector<std::shared_ptr<chrono::ChBody>> bodies_;
-    int num_bodies_;
-    HydroData file_info_;
-    std::vector<ForceFunc6d> force_per_body_;
-    std::shared_ptr<WaveBase> user_waves_;
-
-    // Force components vectors
-    std::vector<double> force_hydrostatic_;
-    std::vector<double> force_radiation_damping_;
-    Eigen::VectorXd force_waves_;
-    std::vector<double> total_force_;  // Saved force per timestep to reduce redundant calculations
-
-    // Additional properties related to equilibrium and hydrodynamics
-    std::vector<double> equilibrium_;
-    std::vector<double> cb_minus_cg_;
-    Eigen::VectorXd rirf_time_vector;  // Assumed consistent for each body
-    Eigen::VectorXd rirf_width_vector;
-
-    // Time tracking for force caching
-    double prev_time;
-
-    // Cached SystemState: built once per time step and reused by all force computations
-    hydrochrono::hydro::SystemState cached_state_;
-    double cached_state_time_ = std::numeric_limits<double>::quiet_NaN();
-
-    // Added mass related properties
-    std::shared_ptr<chrono::ChLoadContainer> my_loadcontainer;
-    std::shared_ptr<ChLoadAddedMass> my_loadbodyinertia;
-
-    /**
-     * @brief Returns the cached SystemState for the given time.
-     *
-     * If cached_state_time_ matches the requested time, returns the cached state.
-     * Otherwise, builds a fresh state and updates the cache (should not happen
-     * in normal flow since CoordinateFuncForBody builds it first).
-     *
-     * @param time Simulation time to get state for
-     * @return Reference to the cached SystemState
-     */
-    const hydrochrono::hydro::SystemState& GetCachedSystemState(double time);
-
-    // Hydrodynamics profiling data (accumulated over run)
-    HydroProfileStats profile_stats_;
-
-    // Hydrostatics force component
-    std::unique_ptr<hydrochrono::hydro::HydrostaticsComponent> hydrostatics_component_;
-
-    // Radiation damping force component
-    std::unique_ptr<hydrochrono::hydro::RadiationComponent> radiation_component_;
-
-    // Wave excitation force component
-    std::unique_ptr<hydrochrono::hydro::ExcitationComponent> excitation_component_;
-
-    // HydroSystem + ChronoHydroCoupler (persistent, constructed once)
-    std::unique_ptr<hydrochrono::hydro::HydroSystem> hydro_system_;
-    std::unique_ptr<hydrochrono::hydro::ChronoHydroCoupler> chrono_coupler_;
-
-    // Legacy compare mode flag: retained for API compatibility; no longer used.
-    bool compare_mode_ = false;
-
-    // Profiling enable flag: stored here because chrono_coupler_ is created lazily.
-    bool profiling_enabled_ = false;
-
-    // Convolution kernel preprocessing (optional)
-    // Uses canonical types from radiation module (single source of truth)
-    hydrochrono::hydro::RadiationConvolutionMode convolution_mode_;
-    bool rirf_processed_ready_ = false;
-    std::vector<Eigen::Tensor<double, 3>> rirf_processed_; // per body [dof x col x step]
-    hydrochrono::hydro::TaperedDirectOptions tapered_opts_;
-    std::string diagnostics_output_dir_;
-
-    void EnsureProcessedRIRF();
-
-    // Helper to ensure radiation component exists with current settings
-    void EnsureRadiationComponent();
-    // Helper to invalidate radiation component (requires full type definition)
-    void InvalidateRadiationComponent();
-    // Helper to ensure excitation component exists
-    void EnsureExcitationComponent();
-
-    // Factory: creates ExcitationComponent with current wave configuration.
-    // Used by both EnsureExcitationComponent() and EnsureHydroSystemAndCoupler()
-    // to ensure consistent construction.
-    std::unique_ptr<hydrochrono::hydro::ExcitationComponent> CreateExcitationComponent() const;
-
-    // Factory: creates HydrostaticsComponent with current equilibrium/stiffness data.
-    // Used by both the HydroForces constructor and EnsureHydroSystemAndCoupler()
-    // to ensure consistent construction.
-    std::unique_ptr<hydrochrono::hydro::HydrostaticsComponent> CreateHydrostaticsComponent() const;
-
-    // Factory: creates RadiationComponent with current BEM data and convolution settings.
-    // Used by both EnsureRadiationComponent() and EnsureHydroSystemAndCoupler()
-    // to ensure consistent construction.
-    // Note: Each instance owns its own velocity history (they are independent).
-    std::unique_ptr<hydrochrono::hydro::RadiationComponent> CreateRadiationComponent() const;
-
-    // Internal helper: constructs hydro_system_ and chrono_coupler_ once.
-    // Subsequent calls are no-ops. Called automatically by CoordinateFuncForBody().
-    void EnsureHydroSystemAndCoupler();
-};
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Backwards-Compatibility Alias
-// ─────────────────────────────────────────────────────────────────────────────
-/**
- * @brief Backwards-compatibility alias for HydroForces.
- *
- * TestHydro is the historical name for the main hydrodynamics façade.
- * It is retained as an alias for backward compatibility with existing code.
- *
- * @note Prefer `HydroForces` for new code.
- *
- * @deprecated Use HydroForces instead.
- */
-using TestHydro = HydroForces;
-
-#endif
