+++ src/hydro/config/setup_from_yaml.h
@@ -1,9 +1,9 @@
 /*********************************************************************
  * @file  setup_from_yaml.h
- * @brief Factory function to create HydroForces from YAML configuration.
+ * @brief Factory function to create HydroSystem from YAML configuration.
  *
  * Provides SetupHydroFromYAML() which matches YAML body configs to Chrono
- * bodies, creates wave models, and returns a configured HydroForces.
+ * bodies, creates wave models, and returns a configured HydroSystem.
  *********************************************************************/
 
 #ifndef HYDRO_CONFIG_SETUP_FROM_YAML_H
@@ -16,24 +16,24 @@
 #include <string>
 
 // Forward declarations
-class HydroForces;
+class HydroSystem;
 
 /**
  * @brief Setup hydrodynamic forces from parsed YAML data.
  *
  * This function connects parsed YAML data (from hydro.yaml) to actual hydrodynamic forces.
  * It builds the appropriate WaveBase subclass from hydro_data.waves, matches body names
- * with their corresponding HDF5 files, and initializes HydroForces with the matched bodies.
+ * with their corresponding HDF5 files, and initializes HydroSystem with the matched bodies.
  *
  * @param hydro_data Parsed hydrodynamic configuration from hydro.yaml
  * @param bodies Vector of Chrono bodies that may have hydrodynamic forces
  * @param timestep Simulation timestep (used for irregular wave setup)
  * @param sim_duration Simulation duration (used for irregular wave setup)
  * @param ramp_duration Wave ramp duration (used for irregular wave setup)
- * @return Unique pointer to initialized HydroForces object
+ * @return Unique pointer to initialized HydroSystem object
  * @throws std::runtime_error on configuration errors
  */
-std::unique_ptr<HydroForces> SetupHydroFromYAML(
+std::unique_ptr<HydroSystem> SetupHydroFromYAML(
     const YAMLHydroData& hydro_data,
     const std::vector<std::shared_ptr<chrono::ChBody>>& bodies,
     double timestep,
@@ -42,7 +42,7 @@ std::unique_ptr<HydroForces> SetupHydroFromYAML(
 );
 
 /**
- * @brief Convenience overload that loads the YAML file and then sets up HydroForces.
+ * @brief Convenience overload that loads the YAML file and then sets up HydroSystem.
  *
  * Behaviour matches calling LoadHydroConfigFromYaml followed by SetupHydroFromYAML.
  *
@@ -51,9 +51,9 @@ std::unique_ptr<HydroForces> SetupHydroFromYAML(
  * @param timestep Simulation time step (used for irregular waves).
  * @param sim_duration Simulation duration (irregular waves).
  * @param ramp_duration Wave ramp duration.
- * @return Initialized HydroForces object.
+ * @return Initialized HydroSystem object.
  */
-std::unique_ptr<HydroForces> SetupHydroFromYAMLFile(
+std::unique_ptr<HydroSystem> SetupHydroFromYAMLFile(
     const std::string& hydro_yaml_path,
     const std::vector<std::shared_ptr<chrono::ChBody>>& bodies,
     double timestep,
@@ -62,4 +62,3 @@ std::unique_ptr<HydroForces> SetupHydroFromYAMLFile(
 );
 
 #endif  // HYDRO_CONFIG_SETUP_FROM_YAML_H
-
